miscellaneous-scripts/create_tasks_precision/requirements.txt
aiofiles==23.1.0
boto3==1.17.49
botocore==1.20.112
certifi==2023.5.7
charset-normalizer==3.2.0
glob2==0.7
idna==3.4
jmespath==0.10.0
psycopg2-binary==2.9.3
pydash==7.0.5
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.3.7
six==1.16.0
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==1.26.16
pandas==2.0.3
